If proc dies between check and sig, do not PANIC
Followed this stacktrace I had:

```python
PANIC! Exception in child worker: 'NoneType' object has no attribute 'send_signal'
Traceback (most recent call last):
  File "/net/storage001.ib.cluster/om2/user/mathsm/ec/enumeration.py", line 205, in _f
    r = f(*a,**k)
  File "/net/storage001.ib.cluster/om2/user/mathsm/ec/enumeration.py", line 249, in solveForTask_ocaml
    timeout=max(int(timeout + 0.5), 1))
  File "/net/storage001.ib.cluster/om2/user/mathsm/ec/enumeration.py", line 38, in run
    self.process.send_signal(signal.SIGUSR1)
AttributeError: 'NoneType' object has no attribute 'send_signal
```

Hypothesis: process exited already thus signal fails, but it exited between the
join and the "send signal" - race condition. The fix can't hurt and I need OM
to stress test so I'm pushing this.

+++ enumeration.py
@@ -34,8 +34,12 @@ class Command(object):
         thread.start()
         thread.join(timeout)   # Wait for finish in less than timeout
                                # If not ready yet, then:
-        if thread.is_alive():  # Ask him nicely to stop, then wait again
-            self.process.send_signal(signal.SIGUSR1)
+                               # Ask him nicely to stop, then wait again
+        if thread.is_alive() and self.process is not None:
+            try:
+                self.process.send_signal(signal.SIGUSR1)
+            except AttributeError:
+                eprint("A process was 'None'. Investiagate")
             thread.join()
 
         return (self.r,self.e)
